refactor: Add deleteUnsupportedEntityTypes() in utils
