+++ src/utils/index.ts
@@ -2,5 +2,6 @@ export { ApiCache } from './api-cache'
 export * from './clear-dead-uuids'
 export * from './create-migration'
 export * from './database'
+export * from './delete-utils'
 export * from './edtr-io'
 export * from './sql-utils'
